Adjust rules to introduce an assembler limit per row
Also don't make lanes containing stone or stone-brick anymore.
Treat any recipes that use those as exclusive (one per row) and
spread them out between all lanes.

This is so that I can place those assemblers right on the edge of the
row, next to the bus, and it can just pull those ingredients in from
the side, unlike every other assembler which will only be able to reach
the normal ingredient lanes for the row.